Features/zk/web service recharge

+++ ttsd-user-repository/src/main/java/com/tuotiansudai/repository/mapper/BankRechargeMapper.java
@@ -13,6 +13,6 @@ public interface BankRechargeMapper {
 
     BankRechargeModel findById(@Param(value = "id") long id);
 
-    long sumRechargeSuccessAmountByLoginName(@Param(value = "loginName") String loginName);
+    int sumRechargeSuccessAmountByLoginName(@Param(value = "loginName") String loginName);
 
 }
